feat: graduate the garbage collection feature
This moves the "garbage collection" feature out of experimental and to
be enabled by default.
