--- test/e2e/suites/gc_managed/gc_managed_suite_test.go
@@ -1,77 +1,0 @@
-//go:build e2e
-// +build e2e
-
-/*
-Copyright 2022 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-	http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package gc_managed //nolint:stylecheck
-
-import (
-	"testing"
-
-	"github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/klog/v2"
-	ctrl "sigs.k8s.io/controller-runtime"
-
-	ekscontrolplanev1 "sigs.k8s.io/cluster-api-provider-aws/v2/controlplane/eks/api/v1beta2"
-	expinfrav1 "sigs.k8s.io/cluster-api-provider-aws/v2/exp/api/v1beta2"
-	"sigs.k8s.io/cluster-api-provider-aws/v2/test/e2e/shared"
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
-	expclusterv1 "sigs.k8s.io/cluster-api/exp/api/v1beta1"
-)
-
-var (
-	e2eCtx *shared.E2EContext
-)
-
-func init() {
-	e2eCtx = shared.NewE2EContext(shared.WithManaged(), shared.WithSchemeInit(initScheme))
-
-	shared.CreateDefaultFlags(e2eCtx)
-}
-
-func TestE2E(t *testing.T) {
-	ctrl.SetLogger(klog.Background())
-	RegisterFailHandler(ginkgo.Fail)
-	ginkgo.RunSpecs(t, "capa-eks-gc-e2e")
-}
-
-var _ = ginkgo.SynchronizedBeforeSuite(func() []byte {
-	return shared.Node1BeforeSuite(e2eCtx)
-}, func(data []byte) {
-	shared.AllNodesBeforeSuite(e2eCtx, data)
-})
-
-var _ = ginkgo.SynchronizedAfterSuite(
-	func() {
-		shared.AllNodesAfterSuite(e2eCtx)
-	},
-	func() {
-		shared.Node1AfterSuite(e2eCtx)
-	},
-)
-
-func initScheme() *runtime.Scheme {
-	sc := shared.DefaultScheme()
-	_ = expinfrav1.AddToScheme(sc)
-	_ = clusterv1.AddToScheme(sc)
-	_ = ekscontrolplanev1.AddToScheme(sc)
-	_ = expclusterv1.AddToScheme(sc)
-
-	return sc
-}
