+++ test/e2e/suites/managed/eks_gc_test.go
@@ -17,58 +17,60 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package gc_unmanaged //nolint:stylecheck
+package managed
 
 import (
 	"context"
 	"fmt"
 	"os"
-	"path/filepath"
 
-	"github.com/gofrs/flock"
 	"github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"k8s.io/utils/ptr"
+	corev1 "k8s.io/api/core/v1"
 
 	infrav1 "sigs.k8s.io/cluster-api-provider-aws/v2/api/v1beta2"
+	ekscontrolplanev1 "sigs.k8s.io/cluster-api-provider-aws/v2/controlplane/eks/api/v1beta2"
 	"sigs.k8s.io/cluster-api-provider-aws/v2/test/e2e/shared"
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
-	controlplanev1 "sigs.k8s.io/cluster-api/controlplane/kubeadm/api/v1beta1"
 	"sigs.k8s.io/cluster-api/test/framework"
-	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
 	"sigs.k8s.io/cluster-api/util"
 )
 
-var _ = ginkgo.Context("[unmanaged] [gc]", func() {
+// EKS cluster external resource gc tests.
+var _ = ginkgo.Describe("[managed] [gc] EKS Cluster external resource GC tests", func() {
 	var (
-		ctx               context.Context
-		result            *clusterctl.ApplyClusterTemplateAndWaitResult
-		requiredResources *shared.TestResource
+		namespace   *corev1.Namespace
+		ctx         context.Context
+		specName    = "eks-extresgc"
+		clusterName string
 	)
 
-	ginkgo.It("[unmanaged] [gc] should cleanup a cluster that has ELB/NLB load balancers", func() {
+	ginkgo.It("[managed] [gc] should cleanup a cluster that has ELB/NLB load balancers", func() {
 		ginkgo.By("should have a valid test configuration")
-		specName := "unmanaged-gc-cluster"
-
-		ctx = context.TODO()
-		result = &clusterctl.ApplyClusterTemplateAndWaitResult{}
-
 		Expect(e2eCtx.Environment.BootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. BootstrapClusterProxy can't be nil")
 		Expect(e2eCtx.E2EConfig).ToNot(BeNil(), "Invalid argument. e2eConfig can't be nil when calling %s spec", specName)
 		Expect(e2eCtx.E2EConfig.Variables).To(HaveKey(shared.KubernetesVersion))
 
-		requiredResources = &shared.TestResource{EC2Normal: 2 * e2eCtx.Settings.InstanceVCPU, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 1, EventBridgeRules: 50}
-		requiredResources.WriteRequestedResources(e2eCtx, specName)
-		Expect(shared.AcquireResources(requiredResources, ginkgo.GinkgoParallelProcess(), flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
-		defer shared.ReleaseResources(requiredResources, ginkgo.GinkgoParallelProcess(), flock.New(shared.ResourceQuotaFilePath))
-		namespace := shared.SetupNamespace(ctx, specName, e2eCtx)
-		defer shared.DumpSpecResourcesAndCleanup(ctx, "", namespace, e2eCtx)
-		ginkgo.By("Creating cluster with single control plane")
-		clusterName := fmt.Sprintf("%s-%s", specName, util.RandomString(6))
+		ctx = context.TODO()
+		namespace = shared.SetupSpecNamespace(ctx, specName, e2eCtx)
+		clusterName = fmt.Sprintf("%s-%s", specName, util.RandomString(6))
 
-		configCluster := defaultConfigCluster(clusterName, namespace.Name)
-		configCluster.WorkerMachineCount = ptr.To[int64](1)
-		createCluster(ctx, configCluster, result)
+		ginkgo.By("default iam role should exist")
+		VerifyRoleExistsAndOwned(ekscontrolplanev1.DefaultEKSControlPlaneRole, clusterName, false, e2eCtx.BootstrapUserAWSSession)
+
+		ginkgo.By("should create an EKS control plane")
+		ManagedClusterSpec(ctx, func() ManagedClusterSpecInput {
+			return ManagedClusterSpecInput{
+				E2EConfig:                e2eCtx.E2EConfig,
+				ConfigClusterFn:          defaultConfigCluster,
+				BootstrapClusterProxy:    e2eCtx.Environment.BootstrapClusterProxy,
+				AWSSession:               e2eCtx.BootstrapUserAWSSession,
+				Namespace:                namespace,
+				ClusterName:              clusterName,
+				Flavour:                  EKSManagedPoolFlavor,
+				ControlPlaneMachineCount: 1, // NOTE: this cannot be zero as clusterctl returns an error
+				WorkerMachineCount:       1,
+			}
+		})
 
 		ginkgo.By(fmt.Sprintf("getting cluster with name %s", clusterName))
 		cluster := framework.GetClusterByName(ctx, framework.GetClusterByNameInput{
@@ -77,6 +79,21 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			Name:      clusterName,
 		})
 		Expect(cluster).NotTo(BeNil(), "couldn't find cluster")
+
+		ginkgo.By("getting AWSManagedControlPlane")
+		cp := GetControlPlaneByName(ctx, GetControlPlaneByNameInput{
+			Getter:    e2eCtx.Environment.BootstrapClusterProxy.GetClient(),
+			Namespace: cluster.Spec.InfrastructureRef.Namespace,
+			Name:      cluster.Spec.ControlPlaneRef.Name,
+		})
+
+		ginkgo.By("Waiting for the machine pool to be running")
+		mp := framework.DiscoveryAndWaitForMachinePools(ctx, framework.DiscoveryAndWaitForMachinePoolsInput{
+			Lister:  e2eCtx.Environment.BootstrapClusterProxy.GetClient(),
+			Getter:  e2eCtx.Environment.BootstrapClusterProxy.GetClient(),
+			Cluster: cluster,
+		}, e2eCtx.E2EConfig.GetIntervals("", "wait-worker-nodes")...)
+		Expect(len(mp)).To(Equal(1))
 
 		workloadClusterProxy := e2eCtx.Environment.BootstrapClusterProxy.GetWorkloadCluster(ctx, cluster.Namespace, cluster.Name)
 		workloadYamlPath := e2eCtx.E2EConfig.GetVariable(shared.GcWorkloadPath)
@@ -97,14 +114,14 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-nlb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeNLB,
 		}, e2eCtx.E2EConfig.GetIntervals("", "wait-loadbalancer-ready")...)
 		shared.WaitForLoadBalancerToExistForService(shared.WaitForLoadBalancerToExistForServiceInput{
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-elb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeELB,
 		}, e2eCtx.E2EConfig.GetIntervals("", "wait-loadbalancer-ready")...)
 
@@ -123,7 +140,7 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-nlb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeNLB,
 		})
 		Expect(err).NotTo(HaveOccurred())
@@ -132,24 +149,20 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-elb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeELB,
 		})
 		Expect(err).NotTo(HaveOccurred())
 		Expect(arns).To(BeEmpty(), "there are %d service load balancers (elb) still", len(arns))
 	})
 
-	ginkgo.It("[unmanaged] [gc] should cleanup a cluster that has ELB/NLB load balancers using AlternativeGCStrategy", func() {
+	ginkgo.It("[managed] [gc] should cleanup a cluster that has ELB/NLB load balancers using AlternativeGCStrategy", func() {
 		ginkgo.By("should have a valid test configuration")
-		specName := "unmanaged-gc-alterstrategy-cluster"
-
-		ctx = context.TODO()
-		result = &clusterctl.ApplyClusterTemplateAndWaitResult{}
-
 		Expect(e2eCtx.Environment.BootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. BootstrapClusterProxy can't be nil")
 		Expect(e2eCtx.E2EConfig).ToNot(BeNil(), "Invalid argument. e2eConfig can't be nil when calling %s spec", specName)
 		Expect(e2eCtx.E2EConfig.Variables).To(HaveKey(shared.KubernetesVersion))
 
+		ctx = context.TODO()
 		shared.ReconfigureDeployment(ctx, shared.ReconfigureDeploymentInput{
 			Getter:       e2eCtx.Environment.BootstrapClusterProxy.GetClient(),
 			ClientSet:    e2eCtx.Environment.BootstrapClusterProxy.GetClientSet(),
@@ -158,21 +171,27 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			WaitInterval: e2eCtx.E2EConfig.GetIntervals("", "wait-deployment-ready"),
 		}, shared.EnableAlternativeGCStrategy, shared.ValidateAlternativeGCStrategyEnabled)
 
-		requiredResources = &shared.TestResource{EC2Normal: 2 * e2eCtx.Settings.InstanceVCPU, IGW: 1, NGW: 1, VPC: 1, ClassicLB: 1, EIP: 1, EventBridgeRules: 50}
-		requiredResources.WriteRequestedResources(e2eCtx, specName)
-		Expect(shared.AcquireResources(requiredResources, ginkgo.GinkgoParallelProcess(), flock.New(shared.ResourceQuotaFilePath))).To(Succeed())
-		defer shared.ReleaseResources(requiredResources, ginkgo.GinkgoParallelProcess(), flock.New(shared.ResourceQuotaFilePath))
-		namespace := shared.SetupNamespace(ctx, specName, e2eCtx)
-		defer shared.DumpSpecResourcesAndCleanup(ctx, "", namespace, e2eCtx)
-		ginkgo.By("Creating cluster with single control plane")
-		clusterName := fmt.Sprintf("%s-%s", specName, util.RandomString(6))
+		specName += "-alterstrategy"
+		namespace = shared.SetupSpecNamespace(ctx, specName, e2eCtx)
+		clusterName = fmt.Sprintf("%s-%s", specName, util.RandomString(6))
 
-		configCluster := defaultConfigCluster(clusterName, namespace.Name)
-		configCluster.WorkerMachineCount = ptr.To[int64](1)
-		c, md, cp := createCluster(ctx, configCluster, result)
-		Expect(c).NotTo(BeNil(), "Expecting cluster created")
-		Expect(len(md)).To(Equal(1), "Expecting one MachineDeployment")
-		Expect(cp).NotTo(BeNil(), "Expecting control plane created")
+		ginkgo.By("default iam role should exist")
+		VerifyRoleExistsAndOwned(ekscontrolplanev1.DefaultEKSControlPlaneRole, clusterName, false, e2eCtx.BootstrapUserAWSSession)
+
+		ginkgo.By("should create an EKS control plane")
+		ManagedClusterSpec(ctx, func() ManagedClusterSpecInput {
+			return ManagedClusterSpecInput{
+				E2EConfig:                e2eCtx.E2EConfig,
+				ConfigClusterFn:          defaultConfigCluster,
+				BootstrapClusterProxy:    e2eCtx.Environment.BootstrapClusterProxy,
+				AWSSession:               e2eCtx.BootstrapUserAWSSession,
+				Namespace:                namespace,
+				ClusterName:              clusterName,
+				Flavour:                  EKSManagedPoolFlavor,
+				ControlPlaneMachineCount: 1, // NOTE: this cannot be zero as clusterctl returns an error
+				WorkerMachineCount:       1,
+			}
+		})
 
 		ginkgo.By(fmt.Sprintf("getting cluster with name %s", clusterName))
 		cluster := framework.GetClusterByName(ctx, framework.GetClusterByNameInput{
@@ -181,6 +200,21 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			Name:      clusterName,
 		})
 		Expect(cluster).NotTo(BeNil(), "couldn't find cluster")
+
+		ginkgo.By("getting AWSManagedControlPlane")
+		cp := GetControlPlaneByName(ctx, GetControlPlaneByNameInput{
+			Getter:    e2eCtx.Environment.BootstrapClusterProxy.GetClient(),
+			Namespace: cluster.Spec.InfrastructureRef.Namespace,
+			Name:      cluster.Spec.ControlPlaneRef.Name,
+		})
+
+		ginkgo.By("Waiting for the machine pool to be running")
+		mp := framework.DiscoveryAndWaitForMachinePools(ctx, framework.DiscoveryAndWaitForMachinePoolsInput{
+			Lister:  e2eCtx.Environment.BootstrapClusterProxy.GetClient(),
+			Getter:  e2eCtx.Environment.BootstrapClusterProxy.GetClient(),
+			Cluster: cluster,
+		}, e2eCtx.E2EConfig.GetIntervals("", "wait-worker-nodes")...)
+		Expect(len(mp)).To(Equal(1))
 
 		workloadClusterProxy := e2eCtx.Environment.BootstrapClusterProxy.GetWorkloadCluster(ctx, cluster.Namespace, cluster.Name)
 		workloadYamlPath := e2eCtx.E2EConfig.GetVariable(shared.GcWorkloadPath)
@@ -201,14 +235,14 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-nlb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeNLB,
 		}, e2eCtx.E2EConfig.GetIntervals("", "wait-loadbalancer-ready")...)
 		shared.WaitForLoadBalancerToExistForService(shared.WaitForLoadBalancerToExistForServiceInput{
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-elb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeELB,
 		}, e2eCtx.E2EConfig.GetIntervals("", "wait-loadbalancer-ready")...)
 
@@ -227,7 +261,7 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-nlb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeNLB,
 		})
 		Expect(err).NotTo(HaveOccurred())
@@ -236,7 +270,7 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 			AWSSession:       e2eCtx.BootstrapUserAWSSession,
 			ServiceName:      "podinfo-elb",
 			ServiceNamespace: "default",
-			ClusterName:      clusterName,
+			ClusterName:      cp.Spec.EKSClusterName,
 			Type:             infrav1.LoadBalancerTypeELB,
 		})
 		Expect(err).NotTo(HaveOccurred())
@@ -251,32 +285,3 @@ var _ = ginkgo.Context("[unmanaged] [gc]", func() {
 		}, shared.DisableAlternativeGCStrategy, shared.ValidateAlternativeGCStrategyDisabled)
 	})
 })
-
-// TODO (richardcase): remove this when we merge these tests with the main eks e2e tests.
-func defaultConfigCluster(clusterName, namespace string) clusterctl.ConfigClusterInput {
-	return clusterctl.ConfigClusterInput{
-		LogFolder:                filepath.Join(e2eCtx.Settings.ArtifactFolder, "clusters", e2eCtx.Environment.BootstrapClusterProxy.GetName()),
-		ClusterctlConfigPath:     e2eCtx.Environment.ClusterctlConfigPath,
-		KubeconfigPath:           e2eCtx.Environment.BootstrapClusterProxy.GetKubeconfigPath(),
-		InfrastructureProvider:   clusterctl.DefaultInfrastructureProvider,
-		Flavor:                   clusterctl.DefaultFlavor,
-		Namespace:                namespace,
-		ClusterName:              clusterName,
-		KubernetesVersion:        e2eCtx.E2EConfig.GetVariable(shared.KubernetesVersion),
-		ControlPlaneMachineCount: ptr.To[int64](1),
-		WorkerMachineCount:       ptr.To[int64](0),
-	}
-}
-
-// TODO (richardcase): remove this when we merge these tests with the main eks e2e tests.
-func createCluster(ctx context.Context, configCluster clusterctl.ConfigClusterInput, result *clusterctl.ApplyClusterTemplateAndWaitResult) (*clusterv1.Cluster, []*clusterv1.MachineDeployment, *controlplanev1.KubeadmControlPlane) {
-	clusterctl.ApplyClusterTemplateAndWait(ctx, clusterctl.ApplyClusterTemplateAndWaitInput{
-		ClusterProxy:                 e2eCtx.Environment.BootstrapClusterProxy,
-		ConfigCluster:                configCluster,
-		WaitForClusterIntervals:      e2eCtx.E2EConfig.GetIntervals("", "wait-cluster"),
-		WaitForControlPlaneIntervals: e2eCtx.E2EConfig.GetIntervals("", "wait-control-plane"),
-		WaitForMachineDeployments:    e2eCtx.E2EConfig.GetIntervals("", "wait-worker-nodes"),
-	}, result)
-
-	return result.Cluster, result.MachineDeployments, result.ControlPlane
-}
